ensure auth works

+++ routes/articles.js
@@ -6,8 +6,11 @@ const { check, validationResult } = require('express-validator');
 //Bring in Models
 let Article = require("../models/article");
 
+//User Model
+let User = require("../models/user")
+
 // Add article route (GET)
-router.get('/add', (req, res) => {
+router.get('/add', ensureAuthenticated, (req, res) => {
     res.render('add_article', {
         title: 'Add Article'
     });
@@ -16,7 +19,7 @@ router.get('/add', (req, res) => {
 // Add article route (POST)
 router.post('/add', [
     check('title', 'Title must not be empty').isLength({ min: 1 }),
-    check('author', 'Author must not be empty').isLength({ min: 1 }),
+    //check('author', 'Author must not be empty').isLength({ min: 1 }),
     check('body', 'Body must not be empty').isLength({ min: 1 })
 ], (req, res) => {
 
@@ -30,7 +33,7 @@ router.post('/add', [
     } else {
         let article = new Article();
         article.title = req.body.title;
-        article.author = req.body.author;
+        article.author = req.user._id;
         article.body = req.body.body;
 
         article.save((err) => {
@@ -46,8 +49,13 @@ router.post('/add', [
 });
 
 // Edit route (GET)
-router.get('/edit/:id', (req, res) => {
+router.get('/edit/:id', ensureAuthenticated, (req, res) => {
     Article.findById(req.params.id, (err, article) => {
+        if (article.author != req.user._id) {
+            req.flash("danger", "Not Authorized");
+            res.redirect("/")
+
+        }
         res.render('edit_article', {
             title: 'Edit Article',
             article: article
@@ -76,24 +84,46 @@ router.post('/edit/:id', (req, res) => {
 
 // Delete route
 router.delete('/:id', (req, res) => {
+    if (!req.user._id) {
+        res.status(500).send();
+    }
     let query = { _id: req.params.id }
-
-    console.log(query);
-    Article.deleteOne(query, (err) => {
-        if (err) {
-            console.log(err);
+    Article.findById(req.params.id, function (err, article) {
+        if (article.author != req.user._id) {
+            res.status(500).send();
+        } else {
+            console.log(query);
+            Article.deleteOne(query, (err) => {
+                if (err) {
+                    console.log(err);
+                }
+                res.send('Success');
+            });
         }
-        res.send('Success');
-    });
+    })
+
 });
 
 // Get single Article
 router.get('/:id', function (req, res) {
     Article.findById(req.params.id, function (err, article) {
-        res.render("article", {
-            article: article
+        User.findById(article.author, function (err, user) {
+            res.render("article", {
+                article: article,
+                author: user.name
+            })
         })
+
     })
 })
 
+//Access Control
+function ensureAuthenticated(req, res, next) {
+    if (req.isAuthenticated()) {
+        return next();
+    } else {
+        req.flash("danger", "Please login");
+        res.redirect("/user/login")
+    }
+}
 module.exports = router;
